new window feature in preview module
